refactor: split technology slot card header and body parts (#3878)
Extract slot header/actions and empty/assigned slot bodies into
library parts; document paths in repo-and-packages.

Refs #3878
